feat: Banco CRUD

+++ database/schema.sql
@@ -17,7 +17,7 @@ CREATE TABLE usuario(
 CREATE TABLE banco(
     id SMALLINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
     nome VARCHAR(45) NOT NULL,
-    saldo_em_centavos INT NOT NULL,
+    saldo_em_centavos BIGINT NOT NULL,
     data_criacao DATETIME NOT NULL DEFAULT NOW(),
     data_edicao DATETIME ON UPDATE NOW(),
     `enabled` TINYINT NOT NULL DEFAULT 1
